01Project_FlowerDeliveryServiceBasic_10/requirements.txt
anyio==4.6.1
asgiref==3.8.1
certifi==2024.8.30
Django==5.0.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
python-dotenv==1.0.1
python-telegram-bot==21.6
sniffio==1.3.1
sqlparse==0.5.1
tzdata==2024.2
gunicorn==20.1.0
#psycopg2-binary==2.9.7
